Create a separate job registry and store actions done
