+++ src/main/java/com/minecolonies/coremod/colony/jobs/registry/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * In this package are all classes regarding registering or creation of new job objects.
+ */
+package com.minecolonies.coremod.colony.jobs.registry;
